+++ lib/src/repositories/socket_states.dart
@@ -1,0 +1,17 @@
+/// Client socket states
+enum SocketStates {
+  /// @nodoc
+  connecting,
+
+  /// @nodoc
+  open,
+
+  /// @nodoc
+  closing,
+
+  /// @nodoc
+  closed,
+
+  /// @nodoc
+  disconnected
+}