+++ lib/xelis_dart_sdk.dart
@@ -2,7 +2,10 @@
 library xelis_dart_sdk;
 
 export 'src/data_transfer_objects/dtos.dart';
-export 'src/repositories/daemon/daemon_channel_repository.dart';
-export 'src/repositories/daemon/daemon_client_repository.dart';
+export 'src/repositories/daemon_client_repository.dart';
+export 'src/repositories/daemon_constants.dart';
+export 'src/repositories/daemon_events_extension.dart';
+export 'src/repositories/daemon_rpc_methods_extension.dart';
+export 'src/repositories/retry_timer.dart';
 export 'src/utils/config.dart';
 export 'src/utils/data_converter.dart';