+++ test/xelis_dart_sdk_test.dart
@@ -1,15 +1,29 @@
 import 'package:test/test.dart';
+// import 'package:xelis_dart_sdk/xelis_dart_sdk.dart';
 
 void main() {
   group('A group of tests', () {
-    // final awesome = Awesome();
+    // final dr = DaemonClientRepository(
+    //   endPoint: localhostAddress,
+    //   secureWebSocket: false,
+    // );
 
-    setUp(() {
+    setUp(() async {
       // Additional setup goes here.
+      // await dr.connect();
     });
 
-    // test('First Test', () {
-    //   expect(awesome.isAwesome, isTrue);
-    // });
+    test('First Test', () {
+      // dr
+      //   ..onOpen(() {
+      //     print('open callback');
+      //   })
+      //   ..onClose(() {
+      //     print('close callback');
+      //   })
+      //   ..onError((error) {
+      //     print('error callback: $error');
+      //   });
+    });
   });
 }